+++ backend/tests/ops_site/cans/test_controller.py
@@ -17,11 +17,3 @@ def test_CAN_fiscal_year_serializer_has_depth_of_one():
 
 def test_CAN_fiscal_year_serializer_returns_all_models_fields():
     assert CANFiscalYearSerializer.Meta.fields == "__all__"
-
-
-def test_Portfolio_serializer_has_depth_of_one():
-    assert PortfolioSerializer.Meta.depth == 1
-
-
-def test_Portfolio_serializer_returns_all_models_fields():
-    assert PortfolioSerializer.Meta.fields == "__all__"
